backend/requirements.txt
Django
djangorestframework
django-cors-headers
djongo
pymongo
python-decouple